Changed build system; Changed ship data container; Added ship destruction

+++ Assets/Scripts/Ship/IShipDataContainer.cs
@@ -10,11 +10,11 @@ namespace Game
         void AddBlock(ShipBlock block);
         void RemoveBlock(ShipBlock block);
 
-        ShipBlock GetBlock(int x, int y);
-        ShipBlock GetBlockByID(int x, int y, int id);
+        ShipBlock GetBlock(float x, float y);
+        ShipBlock GetBlockByID(float x, float y, int id);
 
         IEnumerable<ShipBlock> GetBlocks();
-        IEnumerable<ShipBlock> GetBlocks(int x, int y);
+        IEnumerable<ShipBlock> GetBlocks(float x, float y);
         IEnumerable<ShipBlock> GetBlocksBytID(int id);
     }
 }
